requirements/common.txt
# SHA1:fcc63e365ed7c2d48cf8634eba2735cd5a0cfa45
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
allure-pytest-bdd==2.13.0
    # via -r requirements/common.in
allure-python-commons==2.13.0
    # via allure-pytest-bdd
attrs==22.2.0
    # via
    #   allure-python-commons
    #   pytest
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
decorator==5.1.1
    # via retry2
gherkin-official==24.1.0
    # via -r requirements/common.in
greenlet==2.0.1
    # via playwright
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
lovely-pytest-docker==0.3.1
    # via -r requirements/common.in
mako==1.2.4
    # via pytest-bdd
markupsafe==2.1.2
    # via mako
packaging==23.0
    # via
    #   pytest
    #   pytest-sugar
parse==1.19.0
    # via
    #   parse-type
    #   pytest-bdd
parse-type==0.6.0
    # via pytest-bdd
playwright==1.31.1
    # via -r requirements/common.in
pluggy==1.0.0
    # via
    #   allure-python-commons
    #   pytest
pyee==9.0.4
    # via playwright
pytest==7.2.2
    # via
    #   -r requirements/common.in
    #   allure-pytest-bdd
    #   lovely-pytest-docker
    #   pytest-bdd
    #   pytest-dotenv
    #   pytest-sugar
    #   pytest-testrail
pytest-bdd==6.1.1
    # via
    #   -r requirements/common.in
    #   allure-pytest-bdd
pytest-dotenv==0.5.2
    # via -r requirements/common.in
pytest-sugar==0.9.6
    # via -r requirements/common.in
pytest-testrail==2.9.0
    # via -r requirements/common.in
python-dateutil==2.8.2
    # via -r requirements/common.in
python-dotenv==1.0.0
    # via pytest-dotenv
requests==2.28.2
    # via
    #   -r requirements/common.in
    #   pytest-testrail
    #   testrail-api
retry2==0.9.5
    # via -r requirements/common.in
six==1.16.0
    # via
    #   lovely-pytest-docker
    #   parse-type
    #   python-dateutil
termcolor==2.2.0
    # via pytest-sugar
testrail-api==1.12.0
    # via -r requirements/common.in
typing-extensions==4.5.0
    # via
    #   pyee
    #   pytest-bdd
urllib3==1.26.14
    # via requests
